feat: The RPM class is created
The class is created to host the method getRpmInfo. This way, it will be
possible to mock the filesystem of file_handler and thus remove the use of
existsRegular function.
